frontdeskagent/requirements.txt
aiohttp
aiortc
av
google-genai
numpy